+++ src/Collection/Collection.php
@@ -7,17 +7,20 @@ use TelegramBot\Api\Types\InputMedia\InputMedia;
 /**
  * Class Collection
  */
-class Collection
+class Collection extends \ArrayObject
 {
-    /**
-     * @var array
-     */
-    protected $items = [];
-
     /**
      * @var int Max items count, if set 0 - unlimited
      */
     protected $maxCount = 0;
+
+    /**
+     * @param CollectionItemInterface[] $items
+     */
+    public function __construct(array $items = [])
+    {
+        parent::__construct($items);
+    }
 
     /**
      * @param CollectionItemInterface $item
@@ -33,12 +36,12 @@ class Collection
         }
 
         if ($key == null) {
-            $this->items[] = $item;
+            $this->append($item);
         } else {
-            if (isset($this->items[$key])) {
+            if ($this->offsetExists($key)) {
                 throw new KeyHasUseException("Key $key already in use.");
             }
-            $this->items[$key] = $item;
+            $this->offsetSet($key, $item);
         }
     }
 
@@ -51,7 +54,7 @@ class Collection
     {
         $this->checkItemKey($key);
 
-        unset($this->items[$key]);
+        $this->offsetUnset($key);
     }
 
     /**
@@ -64,15 +67,7 @@ class Collection
     {
         $this->checkItemKey($key);
 
-        return $this->items[$key];
-    }
-
-    /**
-     * @return int
-     */
-    public function count()
-    {
-        return count($this->items);
+        return $this->offsetGet($key);
     }
 
     /**
@@ -82,7 +77,7 @@ class Collection
     public function toJson($inner = false)
     {
         $output = [];
-        foreach ($this->items as $item) {
+        foreach ($this as $item) {
             $output[] = $item->toJson(true);
         }
 
@@ -104,7 +99,7 @@ class Collection
      */
     private function checkItemKey($key)
     {
-        if (!isset($this->items[$key])) {
+        if (!$this->offsetExists($key)) {
             throw new KeyInvalidException("Invalid key $key.");
         }
     }